feat: fix circural reference issue
